+++ frb_rust/src/for_generated/misc_rust_async.rs
